Removes unused dialer mock
We do that by moving dialer related code and tests into separate files
